backend/requirements.txt
# Nexus AI Search Engine Dependencies

# Web Framework
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.0

# HTTP client
aiohttp>=3.8.5
requests>=2.31.0
httpx>=0.25.0

# HTML parsing
beautifulsoup4>=4.12.2
lxml>=4.9.3

# AI and NLP
openai>=1.5.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-anthropic>=0.1.1
langchain-google-genai>=0.0.3
tenacity>=8.2.3

# Vector Databases (Optional: uncomment what you need)
#pinecone-client>=2.2.3
pinecone>=3.0.0
# qdrant-client>=1.6.4
weaviate-client>=3.24.1
# faiss-cpu>=1.8.0

# Web Search
serpapi>=0.1.0
duckduckgo-search>=3.9.0

# Utilities
python-multipart>=0.0.6
ujson>=5.8.0
aiocache>=0.12.1
cachetools>=5.3.1
python-slugify>=8.0.1
colorama>=0.4.6
dataclasses>=0.6
typing-extensions>=4.0.0

# Testing
pytest>=7.4.2
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Performance
orjson>=3.9.5
aiodns>=3.0.0  # For better async DNS resolution
brotli>=1.0.9  # For compression support
cchardet>=2.1.7  # Fast character encoding detection

# Added grpcio dependency
grpcio>=1.71.0
grpcio-tools>=1.71.0

# Monitoring and observability 
prometheus-client>=0.17.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
